Enhance OpenSearch configuration with custom Jackson deserializers for date types and add necessary dependencies #deploy-dolly-search-service #deploy-dolly-backend
