+++ libs/reactive-core/build.gradle
@@ -29,6 +29,9 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-webflux'
     implementation "net.logstash.logback:logstash-logback-encoder:$versions.logback"
 
+    implementation "org.apache.commons:commons-lang3"
+    implementation 'com.fasterxml.jackson.datatype:jackson-datatype-jsr310'
+
     testImplementation "io.projectreactor:reactor-test:$versions.reactorTest"
     testImplementation "org.springframework.boot:spring-boot-webtestclient"
 }
